# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /wallet:
    get:
      operationId: getWallet
      summary: Get my wallet
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/transactions:
    get:
      operationId: listWalletTransactions
      summary: List my wallet transactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransactionsPage"

  # ─── PROMO CODES ─────────────────────────────────────────────────────────────
...
  /admin/wallet/transactions:
    get:
      operationId: listAllTransactions
      summary: List all wallet transactions (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransactionsPage"
  /admin/wallet/refund:
    post:
      operationId: adminRefund
      summary: Issue a refund (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransaction"

  # ─── DRIVER AUTH ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, name, email, phone, role, walletBalance, isVerified, isBlocked, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, driver, admin]
        walletBalance:
          type: number
        isVerified:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Wallet:
      type: object
      required: [balance, userId]
      properties:
        userId:
          type: integer
        balance:
          type: number
    WalletTransaction:
      type: object
      required: [id, userId, amount, type, description, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [deposit, payment, refund]
        description:
          type: string
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    WalletTransactionsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    RefundInput:
      type: object
      required: [userId, amount, description]
      properties:
        userId:
          type: integer
        amount:
          type: number
        description:
          type: string

    # ─── PROMO CODE ──────────────────────────────────────────────────────────
...